Rearranged folder structure in user api, all other files except main apps are placed in resources folder
